+++ client/src/routes/UiTestRoute.js
@@ -54,7 +54,7 @@ const UiTestRoute = () => {
                     <div>Dismissible modal testing.</div>
                 </UiCol>
             </UiDismissibleModal>
-            <div style={{position: "fixed", zIndex: 1000, left: "15px", top: "50%", width: 150, borderRadius: ".35rem", backgroundColor: "#2d2d2d"}}>
+            <div style={{position: "fixed", zIndex: 1000, left: "15px", top: "50%", width: 150, borderRadius: "var(--border-radius)", backgroundColor: "#2d2d2d"}}>
                 <UiContainer className={"py-2 justify-content-center"}>
                     <div style={{textAlign: "center", marginBottom: ".5rem", color: "white"}}>Debug Card</div>
                     <UiButton label={"Dark Mode"} size={"sm"} className={"my-1"} onClick={() => context.onAppearanceToggle()}>Dark Mode</UiButton>