+++ client/src/ui/UiKeyboardInput.js
@@ -6,7 +6,7 @@ const KeyboardInput = styled.kbd`
   font-size: 87.5%;
   color: white;
   background-color: hsl(210, 11%, 15%);
-  border-radius: .35rem;
+  border-radius: var(--border-radius);
   box-shadow: inset 0 -.1rem 0 hsla(0, 0%, 0%, .25);
   .dark & {
     background-color: var(--dark-quaternary) !important;