+++ client/src/ui/form/UiFormControl.js
@@ -9,6 +9,7 @@ const FormControl = styled(Form.Control)`
   color: hsla(0, 0%, 0%, .6);
   background-color: var(--secondary);
   box-shadow: var(--box-shadow);
+  transition: var(--hover-transition);
   &.darker {
     background-color: var(--background);
   }